# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bakery Store API Documentation",
        "description": "API documentation for Bakery Store E-commerce Platform. This API provides endpoints for managing products, orders, carts, customers, and more.",
        "contact": {
            "name": "Le Minh Hoang",
            "email": "leminhhoang.working@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server (Auto-detected from APP_URL)"
        },
        {
            "url": "https://bepmetayapi-9adx6.ondigitalocean.app",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/vouchers/loadCustomerVoucher": {
            "get": {
                "tags": [
                    "Vouchers"
                ],
                "summary": "Get available vouchers for customer",
                "description": "Get all active vouchers for a specific branch/team, grouped by apply_type (discount/shipping_fee)",
                "operationId": "a1bf2fb23eb9181ed2d248f45bf20e34",
                "parameters": [
                    {
                        "name": "team_id",
                        "in": "query",
                        "description": "Team/Branch ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vouchers retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "discount": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "string",
                                                                "format": "uuid"
                                                            },
                                                            "voucher_code": {
                                                                "type": "string"
                                                            },
                                                            "discount_type": {
                                                                "type": "string",
                                                                "enum": [
                                                                    "percent",
                                                                    "fixed"
                                                                ]
                                                            },
                                                            "discount_amount": {
                                                                "type": "number"
                                                            },
                                                            "discount_percent": {
                                                                "type": "number"
                                                            },
                                                            "minimum": {
                                                                "type": "number"
                                                            },
                                                            "remaining": {
                                                                "type": "integer"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "shipping_fee": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "firebaseAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "firebaseAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter Firebase JWT token"
            }
        }
    },
...
}